feat: make Sidebar and top NavigationBar responsive across breakpoints
To make the sidebar more responsive for mobile and small tablets, I
made the sidebar hidden for screens below the lg: tailwind breakpoint.
Thus for mobile/tablets the sidebar visibility is toggled by tapping a
collapse/uncollapse button fixed to the side of the screen.

I updated the styling for elements, so that they are mobile-first. The
css styling needed for larger screen sizes are now prefixed with
sm:, md:, lg: (tailwind's built-in breakpoints).

I also changed certain elements & components such as the SideBarBox,
such that they no longer used a fixed width. They use w-auto instead so
that their width scales with the sidebar's width. And the sidebar's
width scales with the viewport width (20vw). They still have a fixed max
and min width though.