fix(ui): make black border box expand to fill remaining space in viewport
For certain pages, such as the coach page, and the home page, there
is a lot of extra space at the bottom of the page for larger screens.
This is because the UI components in those pages don't take up enough
vertical space.

This is also an UI issue, because a lot of the positioning
of the elements on those pages, depend on the dark container taking up
at minimum all the remaining space in the viewport.

Thus, I added flex-1 (flex-grow: 1), property to the black border box.